3-camera ok

+++ 3-camera/Camera.cpp
@@ -69,9 +69,9 @@ glm::mat4 Camera::calculateViewMatrix()
 
 void Camera::update()
 {
-	front.x = cos(glm::radians(yaw)) * cos(glm::radians(pitch));
-	front.y = sin(glm::radians(pitch));
-	front.z = sin(glm::radians(yaw)) * cos(glm::radians(pitch));
+	front.x = cosf(glm::radians(yaw)) * cosf(glm::radians(pitch));
+	front.y = sinf(glm::radians(pitch));
+	front.z = sinf(glm::radians(yaw)) * cosf(glm::radians(pitch));
 	front = glm::normalize(front);
 
 	right = glm::normalize(glm::cross(front, worldUp));
